+++ models/node.go
@@ -28,7 +28,7 @@ const (
 type Chain struct {
 	ChainID     string    `bson:"chain_id" json:"chain_id"`
 	ChainName   string    `bson:"chain_name" json:"chain_name"`
-	ChainDomain uint64    `bson:"chain_domain" json:"chain_domain"`
+	ChainDomain uint32    `bson:"chain_domain" json:"chain_domain"`
 	ChainType   ChainType `bson:"chain_type" json:"chain_type"`
 }
